+++ .github/workflows/pytest.yaml
@@ -13,7 +13,7 @@ jobs:
     uses: openclimatefix/.github/.github/workflows/python-test.yml@issue/pip-all
     with:
       # pytest-cov looks at this folder
-      pytest_cov_dir: "neso_solar_consumer"
+      pytest_cov_dir: "solar_consumer"
       os_list: '["ubuntu-latest"]'
       python-version: "['3.11']"
       pytest_numcpus: '1'
