+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "neso_solar_consumer"
+name = "solar_consumer"
 version = "0.1"
 dependencies = [
     "pandas",
