+++ solar_consumer/__init__.py
