+++ solar_consumer/app.py
