+++ solar_consumer/data/fetch_gb_data.py
