+++ solar_consumer/data/fetch_nl_data.py
