+++ solar_consumer/fetch_data.py
