+++ solar_consumer/format_forecast.py
