+++ solar_consumer/save_forecast.py
