+++ tests/test_format_forecast.py
@@ -1,5 +1,5 @@
-from neso_solar_consumer.fetch_data import fetch_data
-from neso_solar_consumer.format_forecast import format_to_forecast_sql
+from solar_consumer.fetch_data import fetch_data
+from solar_consumer.format_forecast import format_to_forecast_sql
 from unittest.mock import patch
 import pandas as pd
