+++ tests/test_real_forecasts.py
@@ -19,9 +19,9 @@ import pytest
 import os
 import pandas as pd
 from nowcasting_datamodel.models import ForecastSQL
-from neso_solar_consumer.fetch_data import fetch_data
-from neso_solar_consumer.format_forecast import format_to_forecast_sql
-from neso_solar_consumer.save_forecast import (
+from solar_consumer.fetch_data import fetch_data
+from solar_consumer.format_forecast import format_to_forecast_sql
+from solar_consumer.save_forecast import (
     save_forecasts_to_csv,
     save_forecasts_to_db,
 )
